Backend/requirements.txt
fastapi==0.119.1
uvicorn==0.38.0
python-dotenv==1.1.1
httpx==0.28.1
openai==2.6.0
pypdf==6.1.3
python-multipart==0.0.20
